requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
cython==0.29.26
    # via sagemath
deprecated==1.2.13
    # via redis
packaging==21.3
    # via redis
pydantic==1.9.0
    # via -r requirements.in
pyparsing==3.0.6
    # via packaging
redis==4.1.0
    # via -r requirements.in
sagemath==1.3.0
    # via -r requirements.in
typing-extensions==4.0.1
    # via pydantic
wrapt==1.13.3
    # via deprecated
